Add basic PacletInfo.wl and move source .wl files into new Kernel subdirectory
This changes this package into a Wolfram Language paclet, which is the more modern
way of structuring functionality which extends the Wolfram System.

The new PacletInfo.wl file declares metadata about this paclet (it's name and version),
and the "Kernel" extension declares that this paclet provides WL source code in a
package named "Grassmann`".

This paclet can be loaded by evaluating:

	PacletDirectoryLoad["/path/to/Grassmann.wl/"];

which makes this paclet visible to the system, but does not load the WL source code.
The WL source code can now be loaded using Get (or Needs):

	Get["Grassmann`"]

This works because the "Context" field of the "Kernel" extension informs Get and Needs
that the "Grassmann`" context is provided by this paclet.