Fix Get["*.wl"] to use Get["Grassman`*`"]
Because this package is now a paclet, Get["Grassmann`"] is the canonical and correct
way to load the package. However, doing Get["*`"] behaves differently from Get["*.wl"],
and in particular means that the Get["*.wl"] directives removed in this commit no longer
work, because their relative file paths do not get resolved.

Replacing these with Get["Grassman`*`"] works because the "Grassmann`" context is known
location declared by the PacletInfo.wl file, and the subcontexts can be resolved relative
to that.

The documentation for the WL FindFile[..] function covers this behavior in more detail.