Add BeginPackage["Grassmann`"] at the top of Grassmann.wl to isolate symbols from this package
Prior to this change, symbols from the Grassmann.wl package were being added to
whatever the current value of $Context happened to be at the time the package was
loaded (typically "Global`").

BeginPackage["Grassmann`"] sets the value of $Context to be "Grassmann`", ensuring
that all symbols created by this package are grouped into one context. This prevents
clashing with symbols the user may have defined in their context, and also makes it
easier to inspect which symbols are part of this package:

	Names["Grassmann`*"]

Before this change:

	In[1] := Context[GrassmannQ]

	Out[1] = "Global`"

after this change:

	In[1] := Context[GrassmannQ]

	Out[1] = "Grassmann`"